+++ application/src/app/login/page.tsx
@@ -1,4 +1,10 @@
+"use client";
+
+import { signIn } from "next-auth/react";
+import { useState } from "react";
+import { useRouter } from "next/navigation";
 import React from 'react';
+
 
 const Header: React.FC = () => {
   return (
@@ -9,17 +15,56 @@ const Header: React.FC = () => {
 };
 
 const LoginForm: React.FC = () => {
+    const [userInfo, setUserInfo] = useState({ login: "", password: "" });
+    const [error, setError] = useState("");
+    const router = useRouter();
+
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    setError(""); // Reset error state
+
+    const result = await signIn("credentials", {
+      redirect: false,
+      username: userInfo.login,
+      password: userInfo.password,
+    });
+
+    if (result?.error) {
+      setError(result.error);
+    } else {
+      router.push("/");
+    }
+  }
+  
   return (
     <div className="flex flex-col ml-5 w-[45%] max-md:ml-0 max-md:w-full">
-      <form className="flex flex-col gap-1.5 items-start self-stretch my-auto -ml-px text-stone-600 max-md:mt-10">
+      <form onSubmit={handleSubmit} className="flex flex-col gap-1.5 items-start self-stretch my-auto -ml-px text-stone-600 max-md:mt-10">
         <h2 className="text-2xl font-bold tracking-[2.25px]">log in</h2>
         <div className="mt-3 text-xs text-stone-400">
           <label htmlFor="name" className="sr-only">Name</label>
-          <input type="text" id="name" className="w-full bg-transparent border-b-2 border-rose-400 border-dashed" aria-label="Name" />
+          <input 
+          type="text" 
+          id="login" 
+          value={userInfo.login}
+          onChange={(e) => setUserInfo({ ...userInfo, login: e.target.value })}
+          className="w-full bg-transparent border-b-2 border-rose-400 border-dashed" 
+          aria-label="login" 
+          autoComplete="off" 
+          required 
+          />
         </div>
         <div className="mt-3 text-xs text-stone-400">
           <label htmlFor="password" className="sr-only">Password</label>
-          <input type="password" id="password" className="w-full bg-transparent border-b-2 border-rose-400 border-dashed" aria-label="Password" />
+          <input 
+          type="password" 
+          id="password"
+          value={userInfo.password}
+          onChange={(e) => setUserInfo({ ...userInfo, password: e.target.value })} 
+          className="w-full bg-transparent border-b-2 border-rose-400 border-dashed" 
+          aria-label="Password" 
+          autoComplete="off" 
+          required />
         </div>
         <a href="/signup" className="mt-3 text-xs tracking-wider">don&apos;t have an account?</a>
         <button type="submit" className="self-center px-4 py-1.5 mt-10 text-sm font-bold text-center text-white drop-shadow-mdç bg-sage rounded-[35px] w-[86px] max-md:mt-10">
